remove testing code left in in error
Intlen was still assuring a positive value, this is expected as referenced in the dev notes. (Though nolonger assured through code is significantly faster when scanning positive numbers)  code will have to assert this with abs on the input to the parameter, which will be significantly faster, keep in mind max negative value for a 16 bit integer shifts to zero. Which may cause bugs if not managed by user code.